Added Mana.java, Order.java and Tester.java
The Mana is a class representing a Falafel dish from Shavah.

The Order is a class representing a Falafel order consisting of "Manot" (dishes, nu) of Falafel.

The Tester demonstrates a few aspects of the aforementioned classes, including expected use according to an order flow, and a couple of edge cases.

Hopefully these classes could be the first stone of a backend data manipulation and retrieval structure.